requirements.txt
alembic==0.9.6
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.2
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
apscheduler
prettytable